# internal/platform/handlers/custom/openapi.yaml
openapi: 3.0.3
info:
  title: PDAccess Core Service - OpenAPI 3.0
  description: |-
    Core services for PDAccess.
  termsOfService: http://www.pdaccess.com/terms/
  contact:
    email: support@pdaccess.com
  version: 0.0.1
...
servers:
  - url: /api/v1/ws
...
paths:
  /service:
    post:
      tags:
        - service
      summary: New Service
      description: Create Service for related types/os/versions
      operationId: newService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateServiceResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /service/index:
    post:
      tags:
        - service
      summary: Search Service
      description: Search service for logged user
      operationId: searchService
      parameters:
        - name: group
          in: query
          description: Group Id
          required: false
          explode: true
          schema:
            type: string
        - name: delete
          in: query
          description: Show deleted Service
          required: false
          explode: true
          schema:
            type: boolean
        - name: limit
          in: query
          description: limit count of results
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          description: offset of the page count
          required: false
          schema:
            type: integer
            format: int32

        - name: startDate
          in: query
          description: start date of log range
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: end date of log range
          required: false
          schema:
            type: string
            format: date-time

        - name: filter
          in: query
          description: Filter string for service data. For example name, ip address or port number
          required: false
          explode: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceSearchResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /service/{service}:
    get:
      tags:
        - service
      summary: Find service by id
      description: Returns a single service info
      operationId: serviceById
      parameters:
        - name: service
          in: path
          description: ID of service to return
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleServiceResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - service
      summary: Update a service
      description: Single service update logic
      operationId: serviceUpdate
      parameters:
        - name: service
          in: path
          description: ID of service that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags:
        - service
      summary: Deletes a service
      description: delete a service using soft delete
      operationId: deleteService
      parameters:
        - name: service
          in: path
          description: service id to delete
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /service/{service}/message:
    get:
      tags:
        - service
      summary: Message of the day, show message for related service.
      description: Returns MOTD
      operationId: serviceActiveMessage
      parameters:
        - name: service
          in: path
          description: ID of service to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /service/{service}/alarm:
    get:
      tags:
        - alarm
      summary: Message of the day, show message for related service.
      description: Returns MOTD
      operationId: serviceMessage
      parameters:
        - name: service
          in: path
          description: ID of service to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /group:
    post:
      tags:
        - group
      summary: New Group
      description: Create Group with parent
      operationId: newGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGroupResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /group/index:
    post:
      tags:
        - group
      summary: Search Group
      description: Search group for logged user
      operationId: searchGroup
      parameters:
        - name: parent
          in: query
          description: Parent Group Id
          required: false
          explode: true
          schema:
            type: string
        - name: delete
          in: query
          description: Show deleted Service
          required: false
          explode: true
          schema:
            type: boolean
        - name: limit
          in: query
          description: limit count of results
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          description: offset of the page count
          required: false
          schema:
            type: integer
            format: int32
        - name: startDate
          in: query
          description: start date of log range
          required: false
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: end date of log range
          required: false
          schema:
            type: string
            format: date-time
        - name: filter
          in: query
          description: Filter string for group data. For example name, ip address or port number
          required: false
          explode: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupSearchResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /group/{groupId}:
    get:
      tags:
        - group
      summary: Get Single Group
      description: Get Single Group Info
      operationId: getConfiguration
      parameters:
        - name: groupId
          in: path
          description: Group Id
          required: true
          explode: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SingleGroupResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags:
        - group
      summary: Deletes a group
      description: delete a group using soft delete
      operationId: deleteGroup
      parameters:
        - name: groupId
          in: path
          description: Group Id
          required: true
          explode: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /group/{group}/member:
    post:
      tags:
        - group
      summary: new group member
      description: add new group member
      operationId: createGroupMember
      parameters:
        - name: group
          in: path
          description: ID of group ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMemberRequest"
      responses:
        "200":
          description: group member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    get:
      tags:
        - group
      summary: get group members
      description: get group members
      operationId: groupMembers
      parameters:
        - name: group
          in: path
          description: ID of group ID
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: limit count of results
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          description: offset of the page count
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: group members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMemebersResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

    delete:
      tags:
        - group
      summary: delete group members
      description: delete group members
      operationId: deleteGroupMembers
      parameters:
        - name: group
          in: path
          description: ID of group ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMemberDeleteRequest"
      responses:
        "200":
          description: group member deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /group/{group}/message:
    post:
      tags:
        - group
      summary: new group message
      description: add new group message
      operationId: createGroupMessage
      parameters:
        - name: group
          in: path
          description: ID of group ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupMessageRequest"
      responses:
        "200":
          description: group member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    get:
      tags:
        - group
      summary: get group messages
      description: get group messages
      operationId: searchMessage
      parameters:
        - name: group
          in: path
          description: ID of group ID
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: limit count of results
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          description: offset of the page count
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: group messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMessagesResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

    delete:
      tags:
        - group
      summary: delete group message
      description: delete group message
      operationId: deleteGroupMessage
      parameters:
        - name: group
          in: path
          description: ID of group ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                example:
                description: group message id
      responses:
        "200":
          description: group message deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /snippet:
    post:
      tags:
        - snippet
      summary: New snippet text
      description: New snippet text
      operationId: newSnippet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetRequest"
      responses:
        "200":
          description: create snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnippetResponse"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    get:
      tags:
        - snippet
      summary: get user snippets
      description: get user snippets
      operationId: userSnippets
      parameters:
        - name: limit
          in: query
          description: limit count of results
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          description: offset of the page count
          required: false
          schema:
            type: integer
            format: int32
        - name: marked
          in: query
          description: filter of snippet with marked
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: snippet response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchSnippetResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /alarm:
    get:
      tags:
        - alarm
      summary: current alarm for user
      description: current alarm for user
      operationId: alarmIndex
      parameters:
        - name: limit
          in: query
          description: limit count of results
          required: false
          schema:
            type: integer
            format: int32
        - name: offset
          in: query
          description: offset of the page count
          required: false
          schema:
            type: integer
            format: int32
        - name: filter
          in: query
          description: filter of alarm
          required: false
          schema:
            type: string
      responses:
        "200":
          description: alarm response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlarmsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
components:
  schemas:
...
    CreateServiceRequest:
      type: object
      description: Service is a main artifact for pdaccess.
      required:
        - name
        - groupId
        - accessProtocol
        - authProtocol
        - vendor
        - host
      properties:
        name:
          type: string
          example: new server
        description:
          type: string
          example: new server description
        groupId:
          type: string
          format: uuid
          example: 4ff2d0ee-5479-433e-84c0-f6f8f44a8f50
        accessProtocol:
          $ref: "#/components/schemas/AccessProtocol"
          example: ssh
        authProtocol:
          $ref: "#/components/schemas/AuthProtocol"
          example: radius
        vendor:
          type: string
          description: Name of the service type
          example: Redhat
        version:
          type: string
          description: Version of the service
          example: 9
        host:
          type: string
          example: demo.pdaccess.com
          description: host parameter can be ip or hostname.
        port:
          type: integer
          format: int32
          example: 22
    CreateServiceResponse:
      type: object
      description: seviceId is used in many places to get service data
      properties:
        serviceId:
          type: string
          format: uuid
          example: 6e896583-cefd-498b-ab52-3099bf706d83
    SingleServiceResponse:
      type: object
      properties:
        serviceId:
          type: string
...
    ServiceSearchResponse:
      type: object
      properties:
        serviceId:
          type: string
        name:
          type: string
        description:
          type: string
...